Update patch set 10
Patch Set 10:

I'm not sure if this change goes in the right direction. The concept of the offset how it is used in Monasca is explained here [1].

"The offset is determined by the ID of the last element in the result list."

If we want to change it here. We would also have to change the pagination links.
Please also note that the resource /v2.0/alarms/{alarm_id}/state-history [2] does not have 'start_time' parameter and uses the same offset clause.
Last but not least. Changing the API here may potentially break other applications using it.

[1] https://github.com/openstack/monasca-api/blob/master/docs/monasca-api-spec.md#offset
[2] https://github.com/openstack/monasca-api/blob/master/docs/monasca-api-spec.md#get-v20alarmsalarm_idstate-history

Patch-set: 10
Label: Verified=0