Update patch set 2
Patch Set 2:

Aside from the fact that the library seems to be totally created and maintained by one guy, it seems like a great idea. Maintaining at least two different database configurations is a lot of work and seems like bad practice. If this can be used mainstream all database configuration, it would be a valuable change in my opinion.

Patch-set: 2