Remove required healthcheck import of cassandra
The healthcheck code did an import of the cassandra driver even
if the metrics database is configured to be influx. The cassandra
driver is always installed on devstack but may not be in
containers where space is more of a premium.

Change the code to do a try, except of the driver and work
correctly if no cassandra driver and metrics database is influx

Change-Id: I4055230eb7c203eb2ca82777647ab3b0c62bc367