Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Not really sure if you want to tackle that as well. But similar code should exists also for the influxdb. Both cassandra and influx are optional dependencies so having just plain ping check for influx is not sufficient. We've failed to notice that till know.

Patch-set: 1
Label: Code-Review=-1