Update patch set 18
Patch Set 18: Code-Review+2

Change works great for me, everything is kept as it was in for other components apart from the fact that monasca libraries (common, statsd) are installed for them in dev mode. For example:

  /opt/monasca-persister/bin/pip freeze | grep monasca-
  -e git://git.openstack.org/openstack/monasca-common.git@9188cfcb349122e9b33f70a95ddc2b5bc52a1112#egg=monasca_common

that is expected behaviour.

monasca-api is also installed as development package and finally there are no problems with local vagrant setup where the state of local repo (/opt/stack/monasca-api) was not reflected in a change installed as python package in virtualenv (or system wide). 

This is very very ok. 
I think that we can merge that and proceed with other monasca components in follow ups.

Patch-set: 18
Label: Code-Review=+2