Update patch set 10
Patch Set 10:

In overall, having it properly done requires adding prometheus_client to GR and it will take some time. Also in => https://review.openstack.org/#/c/485470/ I am planning on adding the instrumenting with prometheus_client as pluggable mechanism as we don't want to drop statsd monitoring and/or we would like to be able to simple use noop plugin if monitoring with neither prometheus_client or monasca-statsd is possible.

These are all rough ideas, so all may change in future. Though I already started to convince my folks about the need of having the monitoring done with prometheus to align with monasca-api and containerized deployment.

Patch-set: 10