Update patch set 1
Patch Set 1:

I also launched ```tox -e docs``` with success and the changelog you try to remove has been properly rendered. The point is that Changelog is ignored in .gitignore, so you don't have it in freshly cloned repo, but on the other hand the file is generated for each tox run therefore it is available for the doc run.

Patch-set: 1