Older versions of Kafka only available from archive.apache.org
The apache.mirrors.tds.net site only has the latest two major
versions of Kafka. As Ocata monasca-api needs Kafka 0.8.1.1, this
causes a 404 for devstack installs of stable/ocata.  Simple solution
is to switch to the archive.apache.org site which has versions back
as far as 0.8.0.

Note this url change may need to be done for other branches
in the future.

Change-Id: I7a6d259ffe15e310dd765327114e370289045fec