determine influxdb version on each request if required
When determining influxdb version, the code relies on results of
SHOW DIAGNOSTICS. It is unknown when version banner appeared in
the output of the query, so in case we can't find it there the
version is assumed to be < 0.10.

However, it is incorrect to assume that any error means that
the version is < 0.10. Sometimes error means that influxdb is not
ready to serve connections yet. Ignoring these errors leads to
incorrect state of the application.

Try determining version of influxdb on application startup.
If it fails, retry the attempt on each request.

Change-Id: I813201e0410150cb9c81117232ac3986423e7503
Story: 2001313
Task: 5867