Update patch set 3
Patch Set 3:

To add a little more background, the API container that we are currently using points to a version that does not include these changes. In order to update the API, we would need to include a MySQL schema change, and that is just too risky right now.

Patch-set: 3