Update patch set 1
Patch Set 1: Code-Review-1

The monasca API needs to be able to ingest data in the past as well as in the future.  The most critical reason is to avoid data loss.  E.G. an old agent that comes back online needs to be able to send the data to the API. An external application wants to send data for threshold management, etc.  Moreover an external system (like a ML prediction) needs to be able to post metrics in the future.

All in all this feature if it gets implemented needs to be disabled by default and with both a higher and lower bound configurable.  It is unlikely that 2 minutes is a the ideal value and thus needs to be configurable.  Idem for two weeks in the past.

I am unsure DST would not affect the boundaries.  Basically during DST, 2 weeks is not equal to 2 x 14 x 24h, ie TWO_WEEKS = 2 * 7 * 24 * 60 * 60 * 1000 is false.

Patch-set: 1
Label: Code-Review=-1