Update patch set 1
Patch Set 1:

@nicolas Regarding the two weeks, a delta of two weeks is computed by the number of seconds between now (in seconds since the epoch) and the timestamp of the metric (also seconds since the epoch). DST has no effect on that delta just on the representation of the timestamps as a Date and Time.

As Andrea has stated, the Java implementation of the API (the original API for Monasca) always had this behavior. It was not added to the Python implementation of the API so many people don't know of the old behavior. Since this is effectively new behavior, I agree it should be configurable with the default being off. The default is off but the deltas need to be configurable as well.

Unfortunately, the most common reason for the timestamps being off is the source system is not configured with NTP. The Threshold Engine uses the timestamps of the metrics to determine  Alarm state and if the time gets off by just a couple of minutes it will end up dropping the metrics so time drift can be a big issue.

Patch-set: 1
Reviewer: Gerrit User 11809 <11809@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0