Update patch set 1
Patch Set 1:

Other potential delays in timestamps are network issues, and split-brain for example.  A client may lose connection for many days.  Is there a requirement that every timestamp source must be synchronized to use Monasca?

Let's discuss the effects on the threshold engine in the IRC channel as  I do not understand the benefits of this change.

In addition to default off and configurable upper and lower bounds, this patch requires test cases.

Patch-set: 1