Update patch set 1
Patch Set 1: Code-Review-1

Andrea, could you please create a story describing the motivation and proposed configuration options?

I agree with Nicolas and Craig that the length of both intervals should be configurable. Also tests should be added.

Would you be available to discuss it in the team meeting on Wednesday? If so, please add it to the agenda [1].

[1] https://etherpad.openstack.org/p/monasca-team-meeting-agenda

Patch-set: 1
Label: Code-Review=-1