Factor out REST utils from monasca-api
The purpose of this change is to make helper methods from the
monasca-api available in the monasca-log-api to facilitate
extending the monasca-log-api to support log queries. Only the
low hanging fruit is addressed in this commit. There are additional
methods which could also be moved, but these would require more
substantial refactoring. These changes can be made in a follow up
commit.

Depends-On: I9831866143df39f083f9c12bc0afb10311573d29
Change-Id: Ia02d510fe53ce5ce1c69e0232c5cc2fc0cefc7f2
Story: 2000934
Task: 14369