Factor out REST utils from monasca-api
The purpose of this change is to make helper methods from the
monasca-api available in the monasca-log-api to facilitate
extending the monasca-log-api to support log queries. Only the low
hanging fruit is addressed in this commit. There are additional
methods which could also be moved, but these would require
more substantial refactoring. These changes can be made in a later
commit, if required.

The alternative is that we copy and paste code from the monasca-api
into the monasca-log-api for actions such as role based authentication
which should be the same across both apis. This does not seem like
a good long term solution.

Depends-On: https://review.openstack.org/563103
Change-Id: Ia02d510fe53ce5ce1c69e0232c5cc2fc0cefc7f2
Story: 2000934
Task: 14369