Update patch set 14
Patch Set 14: Code-Review-1

(2 comments)

Thanks for doing this! It would be nice if we could move some of the helper methods to monasca_common as they can be reused in the monasca_log_api. I've started a change for that here: https://review.openstack.org/#/c/563103/ and here: https://review.openstack.org/#/c/563105

What do you think the best way of proceeding is?

Patch-set: 14