Update patch set 14
Patch Set 14:

> (2 comments)
 > 
 > Thanks for doing this! It would be nice if we could move some of
 > the helper methods to monasca_common as they can be reused in the
 > monasca_log_api. I've started a change for that here:
 > https://review.openstack.org/#/c/563103/ and here:
 > https://review.openstack.org/#/c/563105
 > 
 > What do you think the best way of proceeding is?

I think it's better to implement at first those methods in monasca_common in a separate task, make a new release for monasca_common then remove them from monasca_api. because it is another separate task and if we want to do that here and now, then should wait for your implementation to be completed, merged and released. if those methods be ready in monasca_common then I can use them in monasca_log_api as soon as I start to implement oslo_policy for it and you don't need to re-do.

Patch-set: 14
Label: Verified=0