Update patch set 14
Patch Set 14:

I agree that some methods could be refactored and moved to monasca_common. But we don't have to do it all in one change. I prefer the small steps approach. This change has been up for a while and I would like to have it merged soon.

I would like to go with Amir's approach:

* merge this change
* add common code to monasca-common
* refactor monasca-api to use new methods from monasca-common

Should we discuss it in team meeting?

Patch-set: 14