Update patch set 14
Patch Set 14: Code-Review+1

Sounds reasonable to me. After looking a little deeper into duplicated code across the APIs, I think the problem is quite broad and it would be great to discuss how we should deal with it in general moving forwards. With that concern aside, this gets my vote.

Patch-set: 14
Label: Code-Review=+1