Add "monasca-status upgrade check" for pre-upgrade
Add a tool to run before upgrading a Monasca installation
that will check for potential issues and report on readiness to upgrade.

For Monasca, this initial version effectively does nothing, but gives a
framework for future checks to be included.

In the future, it may be worth refactoring to use the oslo_upgradecheck
library, which has not been worked out yet.

This is a community-wide story for Stein [1].

Story: 2003570

[1] https://governance.openstack.org/tc/goals/stein/upgrade-checkers.html

Change-Id: I799602caa4029d67f59307bf1ca0d1e8f254d415