Update patch set 4
Patch Set 4: Code-Review-1

(2 comments)

Please remove the fingerprint: it causes a hash collision. See inline comments for details on why and how that hash collision comes about.

Also, we should fix revision detection for database revisions that are already stamped with Alembic metadata. See inline comments for a bit of an idea of how we could go about this.

Patch-set: 4
Label: Code-Review=-1