Update patch set 4
Patch Set 4:

Looks good, and nice to see it is changed for docker and devstack. Any actions we should take to warn other distributions of the file name change, or is a release note sufficient?  Actually, I see this commit doesn't include a release note, so I'll suggest we add one.

Patch-set: 4