Update patch set 1
Patch Set 1:

> I haven't had a chance to test Kafka 2.x yet - how much testing has
 > been done? 

Not a lot. I used it locally a bit and the CI passes. But why would we handle kafka version bumps different to all the other 100+ components which are updated on a regular base? If somebody detects major problems with the new version, we can a) fix it or b) revert it. And you are still not enforced to use a newer kafka version. This is just the version that is used by devstack...

Patch-set: 1