Detect missing metric_id before passing through bytearray()
It is possible for a row in Cassandra to have a missing metric_hash
(shows as 'null' in cqlsh). This causes an ugly NoneType error
to be passed up to the user on the command line for
'monaca metric-list'.

Fix is to detect the missing value, log an error, and return the
row with None for the metric[id].

This commit is effectively the same as
commit I e617932c6b12a6cfe441510e120bb77a3470b9cf but had
to be reworked to backport to stable/pike.

Change-Id: I51c935534062f7e18a5609830778052b6cc58adb
Story: 2005305
Task: 30194