Fix alarms get method return 404 not found bug
If a user get alarm that not exist in database,then return a 404
not found error to user (if i use postman get the alarm that not
exist,then a blank page without any message except 404 not found).
This is a very bad design.Any erros or exceptions should not
forward to users.

Change-Id: Ia621e8fc3b05bdc1f0c628057f5dbeedd10c9431
Story: 2005648
Task: 30929