Update patch set 2
Patch Set 2: Code-Review+2

We have been running with this setting for a few weeks now. One of the key benefits is that it reduces memory consumption with high cardinality datasets. In one instance, without this setting, InfluxDB would burn through 250GB of RAM on startup, and then get hit by the OOM killer. With this setting, we can now start that instance without running out of memory.

Patch-set: 2
Label: Code-Review=+2