Update patch set 3
Patch Set 3: Code-Review+1

Thanks, looks good to me. The failing monasca-tempest-python2-influxdb job is probably related to the bug in InfluxDB client and will be fixed in requirements repository [1].

[1] https://review.opendev.org/679608

Do you know if the PDF documents are generated in CI job and can be looked up as the job output?

build-openstack-api-ref job generates an empty page. That's because we still keep the API reference in the legacy markdown document [2]. Could we add a pointer to this document in the api-ref index page [3]? Probably best to do in a separate change.

[2] https://opendev.org/openstack/monasca-api/src/branch/master/docs/monasca-api-spec.md
[3] https://opendev.org/openstack/monasca-api/raw/branch/master/api-ref/source/index.rst

Patch-set: 3
Label: Code-Review=+1