Update patch set 11
Patch Set 11:

So we have a better understanding of the strange behaviour now. Looks like the SHOW TAG KEY/VALUE queries in InfluxDB that are scoped within a time range, the results are accurate to shard duration. E.g. if the retention policy is unbounded, the default shard duration is 1 week. On the other hand, when the retention policy is 2 week, the shard duration appears to be 1 day... which explains why we Influxdb is returning results outside of the time window. My feeling is that this is fine for things like metric name lookup queries which we use to narrow down search results to relevant shards as at present, the query is unbounded and looks in the entire database which results in long running queries.

Patch-set: 11