Update patch set 7
Patch Set 7:

I think this code is not needed when working with the Confluent Kafka client. The messages are added to the queue one by one and processed asynchronously. The queue length is controlled by the client and its maximum size can be configured e.g. with queue.buffering.max.kbytes (should we add Monasca configuration options for that?). Thoughts?

Do you need to support the legacy client?

Patch-set: 7