Update patch set 1
Patch Set 1:

An alternative would be to set `num.partitions=3` in the Kafka config file and rely on auto topic generation to create the topics. I have a slight preference for more than 1 partition per topic, since it's most likely to be deployed in the wild that way. On the other hand it's not like devstack is multi-node. What do you think?

Patch-set: 1