Update patch set 3
Patch Set 3:

Good suggestion, I agree that it is a pain to set up at the moment. I suppose it could also be the Notification service that does the Kafka setup, and the API syncs that from the Notification service somehow, perhaps via the database, or Zookeeper. Since this feature is experimental, my preference would be to run it at scale for a bit, and if it works well, we could implement the changes you suggest to make it more user friendly. What do you think?

Patch-set: 3
Reviewer: Gerrit User 17669 <17669@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0