Update patch set 10
Patch Set 10:

> > Any update on this one?
 > 
 > I'm getting close to finalizing the plugin update, so I guess it'd
 > be more suitable to then test the update as a whole.

I think the ES configuration changes which you've added here are crucial for the stability of the CI jobs and solve the problems we've experienced recently. I've based two other changes on top of this one and they are blocked now.
So, if there are no strong arguments against it I'd like to merge this change as soon as possible and continue the work in the new change.

Patch-set: 10