Update patch set 5
Patch Set 5:

Should we un-pin the Kafka messaging version in this commit, or at least advance it?

Last time I tried this upgrade (Logstash 5) I found that that the default Kafka plugins didn't support the 0.9 messaging version, so I had to downgrade them to the 4.x series [1]. I can no longer find this reference in the 7.x release, so maybe it works, but it seems risky.

The change looks sensible to me apart from that, thanks for your efforts!

[1] https://www.elastic.co/guide/en/logstash/5.6/plugins-inputs-kafka.html

Patch-set: 5
Label: Verified=0