Update patch set 5
Patch Set 5:

(1 comment)

In existing deployments it would require migrating to the new Kafka topic name. Renaming topics is not supported so we'd have to delete and create new one. How do we feel about that?

I understand that the proposed name better describes the purpose of the topic and avoids the conflict with the topic for persisting events collected by events API endpoint.

Patch-set: 5
Label: Verified=0