Update patch set 7
Patch Set 7:

Hmm, I agree the naming should make it clear, and would like to make the topic names clear as to their purpose, maybe alarm_events_topic and openstack_events_topic (for the events api).  But we would need to talk through a few migration scenarios.  If the change happens as part of a disruptive upgrade of Monasca, maybe it would be ok.

Patch-set: 7