Update patch set 3
Patch Set 3:

Pierre, thanks for the fix on patchset 1.
Besides of the upgrade to Focal, the library Falcon was updated in the U-C [1]
This is why the rest of Zuul jobs were failing.

I adapted Monasca-API to be compatible with Falcon 2.0.0, 3.0.0 and 3.0.1 in other change,
then I rebase your change on top of it [2]

[1] https://github.com/openstack/requirements/blob/master/upper-constraints.txt#L457
[2] https://review.opendev.org/c/openstack/monasca-api/+/794765

Patch-set: 3
Reviewer: Gerrit User 28062 <28062@4a232e18-c5a9-48ee-94c0-e04e7cca6543>