[CI] remove devstack-gate requirement
devstack-gate has been retired recently, let's remove it from the
required-projects list as it causes zuul config errors [1].

NOTE(elod.illes): on master branch devstack-gate was removed with a
gate fixing patch [2] that fixes multiple issues. stable/2024.1 was
not created as monasca project was marked as inactive in 2024.1 Caracal
release. So, stable/2023.2 is the first branch where we have to remove
devstack-gate after master.

[1] https://zuul.opendev.org/t/openstack/config-errors?project=openstack%2Fmonasca-api&severity=error
[2] Ie77ccf6b15593520ae4f486843cdc46d7e25c628

Change-Id: Icbd89dcd82a7225918933b09321cbf394f6035d0